Introduce interfaces for ScenarioContext and BuildContext
