+++ src/test/groovy/org/gradle/profiler/mutations/AbstractMutatorTest.groovy
@@ -1,13 +1,13 @@
 package org.gradle.profiler.mutations
 
+import org.gradle.profiler.DefaultScenarioContext
 import org.gradle.profiler.Phase
-import org.gradle.profiler.ScenarioContext
 import org.junit.Rule
 import org.junit.rules.TemporaryFolder
 import spock.lang.Specification
 
 abstract class AbstractMutatorTest extends Specification {
     @Rule TemporaryFolder tmpDir = new TemporaryFolder()
-    def scenarioContext = new ScenarioContext(UUID.fromString("276d92f3-16ac-4064-9a18-5f1dfd67992f"), "testScenario")
+    def scenarioContext = new DefaultScenarioContext(UUID.fromString("276d92f3-16ac-4064-9a18-5f1dfd67992f"), "testScenario")
     def buildContext = scenarioContext.withBuild(Phase.MEASURE, 7)
 }
